isky-data-preprocessing-mservice/requirements.txt
aniso8601==9.0.1
appwrite==1.1.0
async-repool==1.0.1
attrs==21.4.0
bcrypt==3.2.2
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cycler==0.11.0
dependency-injector==4.39.1
Flask==2.1.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.2
Flask-Login==0.6.1
flask-openapi3==1.1.4
flask-restx==0.5.1
fonttools==4.37.1
future==0.18.2
idna==3.3
importlib-metadata==4.11.4
importlib-resources==5.7.1
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.5.1
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.5.3
mistune==2.0.2
numpy==1.23.3
packaging==21.3
pandas==1.4.4
Pillow==9.2.0
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pydantic==1.9.1
PyJWT==2.4.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==7.1.3
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
requests==2.27.1
rethinkdb==2.4.9
RethinkPool==0.4.1
scikit-learn==1.1.2
scipy==1.9.1
seaborn==0.12.0
six==1.16.0
sklearn==0.0
tenacity==8.0.1
threadpoolctl==3.1.0
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
